refactor(mobile): UI parsea y muestra centavos enteros
Sprint flutter-integer-cents-v1, fase 4 de 5 (RF-006, RN-IC-03/04).

**Entrada**: todos los forms parsean con `parseCents` vía
`parseFormattedAmount`, que ahora devuelve `int?` en centavos y traga la
`FormatException` (los forms usan `null` como señal de monto inválido).
Cubre entry_form, account_form, category_form, loan_form, los dos forms
de pago de préstamo, budget_item_form_sheet y el panel de filtros.

**Salida**: `formatAmount(num)` se elimina y `lib/widgets/amount_formatter.dart`
se borra; todo el display pasa por `formatCents(int)`.
`formatAmountForInput` hidrata el TextField desde centavos.

Detalles que necesitaron criterio, no reemplazo mecánico:

- **Slider del split de préstamo**: la API de Material opera en `double`.
  El widget recibe centavos y convierte sólo en la frontera de render;
  al mover, redondea al centavo y deriva el interés por resta exacta
  para que el split siempre cierre contra el total (RN-L08).
- **Reajuste proporcional del split** al cambiar el monto: la ratio sigue
  siendo fraccionaria, pero el capital se redondea y el interés toma el
  resto exacto.
- **TweenAnimationBuilder del subtotal acumulado** (budgets): interpola
  en `double` para que el conteo sea suave y redondea al centavo sólo
  para el display.
- **BudgetProgress / heatmaps**: `usedPct` y los umbrales de cuartil
  siguen en `double` — son porcentajes e interpolaciones, no montos.

+++ mobile/lib/screens/weekly_budgets/detail_screen.dart
@@ -65,10 +65,10 @@ class _WeeklyBudgetScreenState extends State<WeeklyBudgetScreen> {
     _totalsStream = _itemsStream!.map((items) {
       final income = items
           .where((i) => i.kind == 'income')
-          .fold<double>(0, (acc, i) => acc + i.amount);
+          .fold<int>(0, (acc, i) => acc + i.amount);
       final expense = items
           .where((i) => i.kind == 'expense')
-          .fold<double>(0, (acc, i) => acc + i.amount);
+          .fold<int>(0, (acc, i) => acc + i.amount);
       return BudgetTotals(income: income, expense: expense);
     });
     _loadCategories();
